refactor: separate DataBackup and ErrorReporting components, remove the original (large) AccountMenu component
